ARM64: Revive cpu tracking in mm_struct
ARM64 used to have the ability to track which cpu has
used an address space (mm_struct). The cpumask stored in
the mm_struct is currently unused. So what is needed is still
there from the core kernel.

Revert the removal patch and update it to work with the current kernel source.

Signed-off-by: Christoph Lameter (Ampere) <cl@linux.com>